Link entre JavaFx et Spring: Quand on lance HowManagerApplication -> lance javafx, quand on ferme l'interface : -> ferme Spring App.java renommée en JavaFxApplicationSupport.java
