--- src/main/java/com/JESIKOM/HowManager/App.java
@@ -1,31 +1,0 @@
-package com.JESIKOM.HowManager;
-
-import java.io.IOException;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-
-
-public class App extends Application {
-    @Override
-    public void start(Stage primaryStage) throws IOException {
-        FXMLLoader loader = new FXMLLoader(getClass().getResource("/views/main.fxml"));
-
-        Parent root = loader.load();
-
-        primaryStage.setTitle("How Manager");
-        Scene scene = new Scene(root, 950, 700);
-
-        primaryStage.setScene(scene);
-        primaryStage.show();
-
-    }
-
-    public static void main(String[] args) {
-        launch();
-    }
-}
