+++ src/main/java/com/JESIKOM/HowManager/HowManagerApplication.java
@@ -1,13 +1,25 @@
 package com.JESIKOM.HowManager;
 
+import org.springframework.boot.ApplicationRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ConfigurableApplicationContext;
+
+import static javafx.application.Application.launch;
+
 
 @SpringBootApplication
 public class HowManagerApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(HowManagerApplication.class, args);
+		ConfigurableApplicationContext context = SpringApplication.run(HowManagerApplication.class, args);
+
+		// Lancer JavaFX
+		JavaFxApplicationSupport.setContext(context);
+		launch(JavaFxApplicationSupport.class, args);
+
+		// Fermer Spring quand JavaFX se ferme
+		context.close();
 	}
 
 }
